move global types to dedicated file
