+++ src/scenes/keyboard/slice.ts
@@ -1,6 +1,7 @@
 import { createSelector, createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { RootState } from '../../app/store';
-import { CompleteNote, NoteName, ScaleObj, SCALES, ScaleStatus, getOctaveScaleObject, getAllOctaveNotesBetween, convertOctaveNoteToMidiId, getKeyScaleObject } from '../../utils/music';
+import { CompleteNote, NoteName, ScaleObj, ScaleStatus } from '../../types';
+import { SCALES, getOctaveScaleObject, getAllOctaveNotesBetween, convertOctaveNoteToMidiId, getKeyScaleObject } from '../../utils/music';
 
 export interface KeyboardState {
   activeKey: string | null;
